requirements.txt
absl-py==2.1.0
addict==2.4.0
albumentations==1.3.1
astunparse==1.6.3
basicsr==1.4.2
bitsandbytes
beautifulsoup4==4.12.3
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.0
customtkinter==5.2.0
cycler==0.12.1
Cython==3.0.8
darkdetect==0.8.0
diffusers==0.37.0
easydict==1.12
face-alignment==1.4.1
facexlib==0.3.0
fastapi==0.122.0
filelock==3.13.1
filterpy==1.4.5
flatbuffers==24.3.25
fonttools==4.49.0
fsspec==2024.2.0
future==0.18.3
gast==0.4.0
gdown==5.1.0
gfpgan==1.3.8
google-auth==2.28.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
gradio==5.0.0
grpcio==1.60.1
h5py==3.11.0
huggingface_hub==0.36.2
humanfriendly==10.0
idna==3.6
imageio==2.34.0
imgaug==0.4.0
importlib-metadata==7.0.1
insightface==0.7.3
Jinja2==3.1.3
joblib==1.3.2
keras==3.5.0
kiwisolver==1.4.5
kornia==0.8.2
lazy_loader==0.3
libclang==16.0.6
llvmlite==0.44.0
lmdb==1.4.1
Markdown==3.5.2
MarkupSafe==2.1.5
matplotlib==3.8.3
mpmath==1.3.0
networkx==3.2.1
numba==0.61.2
numpy==1.26.4
oauthlib==3.2.2
onnx==1.14.0
onnxruntime-gpu==1.15.1
opencv-python==4.8.0.74
opencv-python-headless==4.9.0.80
opennsfw2==0.10.2
opt-einsum==3.3.0
packaging==23.2
peft==0.18.0
Pillow==10.0.0
platformdirs==4.2.0
prettytable==3.9.0
protobuf==4.23.4
psutil==5.9.5
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.10.6
pyparsing==3.1.1
pyreadline3==3.4.1
PySide6==6.11.0
PySocks==1.7.1
python-dateutil==2.8.2
PyYAML==6.0.1
qudida==0.0.4
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.12.0
shapely==2.0.3
six==1.16.0
soupsieve==2.5
sympy==1.12
tb-nightly==2.17.0a20240216
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.3.0
tifffile==2024.2.12
tk==0.1.0
tkinterdnd2==0.3.0
tomli==2.0.1
torchao==0.15.0
transformers==4.57.6
tqdm==4.65.0
typing-extensions==4.15.0
urllib3==2.2.0
wcwidth==0.2.13
Werkzeug==3.0.1
wrapt==1.16.0
yapf==0.40.2
zipp==3.17.0
